add max order mapping for product flash sale on etalase page

+++ features/merchant/shop_page/src/main/java/com/tokopedia/shop/product/view/datamodel/ShopProductUiModel.kt
@@ -46,7 +46,7 @@ class ShopProductUiModel : BaseShopProductViewModel, ImpressHolder {
     var productInCart: Int = 0
     var isVariant: Boolean = false
     var isNewData: Boolean = false
-    var stock: Int = 0
+    var stock: Long = 0
     var minimumOrder: Int = 0
     var maximumOrder: Int = 0
     var parentId: String = ""
